backend-drf/requirements.txt
django==6.0
python-decouple
djangorestframework
django-cors-headers
djangorestframework-simplejwt
yfinance
pandas
numpy
matplotlib
scikit-learn
tensorflow
keras
mangum
boto3
pillow